backend/loanshark/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==2.4.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
Markdown==3.3.4
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
